pyproject.toml
[project]
name = "repo-digest-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "jinja2>=3.1.4",
    "langchain>=0.3.9",
    "langchain-core>=0.3.21",
    "langchain-openai>=0.2.11",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.8.0",
]
